make object in QueryBuilder not nullable
We might have to discuss, wether we consider this to be a breaking change compared to 3.0.0. I have named the new version 3.0.1 for now. I think this should be fine.

This change makes sense, as setting object to null (possible with the prior default constructor) would brake the whole QueryBuilder.